Set up Native Messaging and reload extensions
Get the extension IDs from the Native Messaging host and reload the
specified extensions. Also reload the current tab.

Turns out this doesn't work the way I expected. It does disable and
re-enable the extension, but it doesn't reload the extension as with
`chrome.runtime.reload()`. This means the specified extension isn't
reloaded with the latest code changes.

Unfortunately, it looks like there's no API to do what I want, and
unless there's some magic in the `chrome.debugger` API I'll have to give
up on this project.