main.lisp: Filter DevTools targets to chosen extensions
Filter a list of extension background page targets.

The extension IDs should come from the command line, but I've hard-coded
the list here.

Increased the `sleep` time to allow time for the messages to be sent &
received before the program exits. Will need to figure out a proper way
to do this later.