main: Ensure WebSocket connection is always closed
Create a new `with-websocket-connection` decorator that starts and
closes a WebSocket connection to the given `client` around the body
forms.

Thanks to Practical Common Lisp's "The Special Operators" chapter
(http://www.gigamonkeys.com/book/the-special-operators.html#unwinding-the-stack)
for introducing me to `unwind-protect`.

Couldn't figure out how to get the new macro to auto-indent properly
with Vlime, so ended up manually indenting it.