main: Move option parsing code to new function `parse-options`
Make a new function `parse-options` that parses the command line options
and returns a `config` object. We'll use that object instead of
`options` in `main`. Cleans up the `main` function a bit.

Currently, we just print the `config` object to ensure we're storing the
proper values.

Followed Practical Common Lisp's example to implement `print-object` so
we can see the contents of its slots:
http://www.gigamonkeys.com/book/practical-a-spam-filter.html#the-heart-of-a-spam-filter

Still need to implement error checking for a missing `--socket-url`
option.