main: Get WebSocket URL from command line for `*client*`
Keep the `*client*` global variable, but use the client in `config`,
constructed from the WebSocket URL passed from the command line.

Not a huge fan of the global variable, but it's way easier to keep it
like this rather than passing the client down to the `attach-to-target`
and `reload-extension` functions.