Make DevTools Protocol call ID auto-incrementing
Remove the hard-coded call IDs and replace them with a class that keeps
track of the current call ID and allows for easy incrementing to get the
next ID.

This should allow us to give multiple extension IDs on the command line
and send messages with properly incrementing call IDs.

Didn't touch the `runtime-evaluate-msg` message call ID because that one
is local to the target it's attached to, so we can keep it at ID "1".