Remove web extension and native messaging host
This code is superseded by the Common Lisp project that communicates via
the DevTools Protocol.

The `chrome.management` API's `setEnabled()` function just allowed me to
turn extensions off and on. It didn't reload the extensions.

The DevTools Protocol allows us to execute JavaScript in the context of
an extension's background page. This allows us to run
`chrome.runtime.reload()` in an extension's context, properly reloading
the extension.