Makefile: Rewrite `pkg` target to include bundle and source
Copy the `pkg` target from Wajir to make a tarball of the bundled
dependencies in `bundle` and the Extreload source code.

This tarball can then be distributed and the program can be built and
installed from it using the Makefile. By distributing the source, an
executable can be built for more architectures and systems.