Track Manifest V3 extensions in a list
I want to keep track of Manifest V3 extensions, because these must be
re-attached to in order to reload the tab. We will later look at this
list to find out if we need to skip reloading the extension (it's
already been done), and instead proceed directly to reloading the tab.