main: Try to force re-attach to service worker to reload the current tab
Doesn't appear to be working, unfortunately. The idea was to force the
execution of the re-attach instead of checking for a targetCrashed
message (which I would get only very rarely). In this way, we ensure
that we re-attach to the service worker before running the tab reload.
But that doesn't seem to be enough.